added readme.md changed protocol models
