+++ OpenImis.RestApi/Properties/PublishProfiles/FolderProfile.pubxml
@@ -18,6 +18,6 @@ by editing this MSBuild file. In order to learn more about this please visit htt
     <SelfContained>true</SelfContained>
     <_IsPortable>false</_IsPortable>
     <publishUrl>bin\Debug\netcoreapp2.1\publish\</publishUrl>
-    <DeleteExistingFiles>False</DeleteExistingFiles>
+    <DeleteExistingFiles>True</DeleteExistingFiles>
   </PropertyGroup>
 </Project>
