+++ OpenImis.RestApi/Protocol/IRequestModel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace OpenImis.RestApi.Models.HTTPModels
+namespace OpenImis.RestApi.Protocol
 {
     interface IRequestModel
     {
